[refactor] Novel controller, api dto -> web 아래로 이동 - #9
