+++ novelservice/src/main/java/com/iucyh/novelservice/novel/service/NovelService.java
@@ -4,12 +4,12 @@ import com.iucyh.novelservice.novel.exception.DuplicateNovelTitle;
 import com.iucyh.novelservice.novel.exception.NovelNotFound;
 import com.iucyh.novelservice.novel.domain.Novel;
 import com.iucyh.novelservice.novel.enumtype.NovelCategory;
-import com.iucyh.novelservice.novel.dto.mapper.NovelRequestMapper;
-import com.iucyh.novelservice.novel.dto.mapper.NovelResponseMapper;
-import com.iucyh.novelservice.novel.dto.request.CreateNovelRequest;
-import com.iucyh.novelservice.novel.dto.response.NovelLikeCountResponse;
-import com.iucyh.novelservice.novel.dto.request.UpdateNovelRequest;
-import com.iucyh.novelservice.novel.dto.response.NovelResponse;
+import com.iucyh.novelservice.novel.web.dto.mapper.NovelRequestMapper;
+import com.iucyh.novelservice.novel.web.dto.mapper.NovelResponseMapper;
+import com.iucyh.novelservice.novel.web.dto.request.CreateNovelRequest;
+import com.iucyh.novelservice.novel.web.dto.response.NovelLikeCountResponse;
+import com.iucyh.novelservice.novel.web.dto.request.UpdateNovelRequest;
+import com.iucyh.novelservice.novel.web.dto.response.NovelResponse;
 import com.iucyh.novelservice.novel.repository.NovelRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
