+++ novelservice/src/main/java/com/iucyh/novelservice/novel/web/dto/response/NovelLikeCountResponse.java
@@ -1,3 +1,3 @@
-package com.iucyh.novelservice.novel.dto.response;
+package com.iucyh.novelservice.novel.web.dto.response;
 
 public record NovelLikeCountResponse(int likeCount) {}
